+++ apps/web/app/app.dub.co/(dashboard)/[slug]/settings/oauth-apps/page-client.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { clientAccessCheck } from "@/lib/api/tokens/permissions";
 import useWorkspace from "@/lib/swr/use-workspace";
 import { OAuthAppProps } from "@/lib/types";
 import OAuthAppCard from "@/ui/oauth-apps/oauth-app-card";
@@ -8,22 +7,15 @@ import OAuthAppPlaceholder from "@/ui/oauth-apps/oauth-app-placeholder";
 import EmptyState from "@/ui/shared/empty-state";
 import { Cube } from "@dub/ui";
 import { fetcher } from "@dub/utils";
-import { useRouter } from "next/navigation";
 import useSWR from "swr";
 
 export default function OAuthAppsPageClient() {
-  const router = useRouter();
-  const { slug, id: workspaceId, role } = useWorkspace();
+  const { id: workspaceId } = useWorkspace();
 
   const { data: oAuthApps, isLoading } = useSWR<OAuthAppProps[]>(
     `/api/oauth/apps?workspaceId=${workspaceId}`,
     fetcher,
   );
-
-  const { error: permissionsError } = clientAccessCheck({
-    action: "oauth_apps.write",
-    role,
-  });
 
   return (
     <div className="grid gap-5">
